+++ .gitignore
@@ -1,2 +1,3 @@
+/js/impress.min.js
 /node_modules
 /npm-debug.log
