+++ qunit_test_runner.html
@@ -20,7 +20,7 @@
   <!-- Core tests -->
   <script src="test/core_tests.js"></script>
   <!-- Plugins -->
-  <script src="test/navigation_tests.js"></script>
+  <script src="src/plugins/navigation/navigation_tests.js"></script>
 
 </body>
 </html>
